+++ lib/provider.dart
@@ -1,8 +1,4 @@
 library provider;
 
-export 'src/async_provider.dart';
-export 'src/consumer.dart';
 export 'src/delegate_widget.dart';
-export 'src/listenable_provider.dart';
-export 'src/provider.dart';
-export 'src/proxy_provider.dart';
+export 'src/provider.dart' hide Disposer;
